Split large file with tests into several smaller ones
