+++ src/test_inject/__init__.py
@@ -1,0 +1,8 @@
+from unittest import TestCase
+
+import inject
+
+
+class BaseTestInject(TestCase):
+    def tearDown(self):
+        inject.clear()
