+++ src/test_inject/test_attr.py
@@ -1,0 +1,27 @@
+import inject
+from test_inject import BaseTestInject
+
+
+class TestInjectAttr(BaseTestInject):
+    def test_attr(self):
+        class MyClass(object):
+            field = inject.attr(int)
+
+        inject.configure(lambda binder: binder.bind(int, 123))
+        my = MyClass()
+        value0 = my.field
+        value1 = my.field
+
+        assert value0 == 123
+        assert value1 == 123
+
+    def test_class_attr(self):
+        class MyClass(object):
+            field = inject.attr(int)
+
+        inject.configure(lambda binder: binder.bind(int, 123))
+        value0 = MyClass.field
+        value1 = MyClass.field
+
+        assert value0 == 123
+        assert value1 == 123
